Get the building style names from the more building styles DLL
This fixes a bug where the optional auto-historical check box would be
shown as a building style in the query dialog.
If the user doesn't have the more building styles DLL installed, the
code will check for the 4 built-in Maxis building styles.

Unknown building styles are ignored because building styles are a
subset of the game's occupant groups, and there is no way to tell
an unknown building style from any other occupant group value.