Add an option to disable the occupant query sounds
These sounds are played when clicking on an occupant to open the query
dialog or copy the lot.
The occupant query sound properties affected by this option are
SFX:Query Sound, SFX:Query Sound Abandoned and SFX:Query Sound Decayed.

With this option active, the game's call to the PlayOccupantQuerySound
function is hooked and redirected to a version that does nothing.

To support this the DLL how includes an INI file.