Fix the UnknownTokenCallback return value
The callback must return true if the token was processed, this tells
the game that it can stop searching for a token handler.

A failure of the method that handles the token now results in an empty
string being displayed in-game.
This prevents the user from seeing the name of the underlying UI label.