Update GetBuildingWallToWallToken
Replaced the cIBuildingStyleWallToWall interface with the
cIBuildingStyleInfo2 interface.
The cIBuildingStyleInfo2::IsWallToWall method is easier to read.

The cIBuildingStyleWallToWall::HasWallToWallOccupantGroup and
cIBuildingStyleInfo2::IsWallToWall methods both check for the
"Building Is Wall-to-Wall" property in addition to the occupant groups,
so the HasWallToWallOccupantGroup name is no longer 100% accurate as to
what it does.