Add error reporting to the building_is_w2w variable
The building_is_w2w variable depends on the MoreBuildingStyle DLL to
provide that information.
If the DLL is not installed or the version is too old, loading the
cIBuildingStyleInfo2 GZCOM class it provides will fail.

When this occurs we write an error message to the log file, and return
a W2W status of Unknown.