Add Shift + Click copying for some flora types
This only works for the flora types that are highlighted when hovering
over them with the query tool.
The other types appear to be static models that are not considered
flora occupants by the game.

The Shift + Click copying is now implemented as a jump in the
cSC4ViewInputControlQuery::OnMouseDownL function, the jump replaces a
disabled debug command that Maxis left in the game code.