Extend the TerrainQuery command output
It now prints the floating point x, y and z coordinates without
truncating to the first decimal place, and the cell information now
includes the moisture value.

Added an advanced/debug version of TerrainQuery that includes the
humidity, ambient wind speed, and ambient wind directions.
These values appear to be the same across a city tile.
The weather simulator doesn't provide a way to get the cell temperature
data, even though the game appears to be calculating it.