Add an optional pollution mode to TerrainQuery
This mode is activated by holding the control key when moving the query
tool.
The TerrainQuery cheat must be active for this to have any effect.

When active, it shows the flammability, land value, and pollution
(air, water, garbage, and radiation).
This data is the same as what the Maxis advanced building query shows,
but works for any land cell.